Passage des questions en QCM : migrations + seeder + factory + front OK To Do: Quizz - Back

+++ database/factories/UserFactory.php
@@ -24,21 +24,25 @@ class UserFactory extends Factory
         $filename = Str::slug($name);
 
         //? Image Avatar
-        Storage::disk('public')->put('img/profile/' . $filename . '.jpg', file_get_contents('https://loremflickr.com/300/300/girl,boy,man,woman'));
+        $imageContent = file_get_contents('https://loremflickr.com/300/300/girl,boy,man,woman');
+        // Si on a bien une image
+        if (!empty($imageContent)) {
+            Storage::disk('public')->put('img/profile/' . $filename . '.jpg', $imageContent);
 
-        // Transformation en avif
-        $gdImage = imagecreatefromjpeg(storage_path('app/public/img/profile/' . $filename . '.jpg'));
-        $resizeBigImg = ImageTransformation::image_resize_small($gdImage, 300, 300);
-        \imageavif($resizeBigImg, storage_path('app/public/img/profile/' . $filename . '.avif'));
+            // Transformation en avif
+            $gdImage = imagecreatefromjpeg(storage_path('app/public/img/profile/' . $filename . '.jpg'));
+            $resizeBigImg = ImageTransformation::image_resize_small($gdImage, 300, 300);
+            \imageavif($resizeBigImg, storage_path('app/public/img/profile/' . $filename . '.avif'));
 
-        imagedestroy($gdImage);
-        imagedestroy($resizeBigImg);
-        // On efface le png original
-        unlink(storage_path('app/public/img/profile/' . $filename . '.jpg'));
+            imagedestroy($gdImage);
+            imagedestroy($resizeBigImg);
+            // On efface le png original
+            unlink(storage_path('app/public/img/profile/' . $filename . '.jpg'));
+        }
 
         return [
             'name' => $name,
-            'avatar' => $filename . '.avif',
+            'avatar' => !empty($imageContent) ?  $filename . '.avif' : null,
             'email' => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
             'password' => bcrypt(123456), // password
